No per commit comments until release to master.
Notes: Added a few things that have been in my TODO list for a while. Still more polishing work for push to master. Have gotten busy again for a few weeks which is why updates have gotten slower.

USE WITH CAUTION, It should work fine, but make a backup before attempting to use.